toml update, sym matrix update
