+++ pyproject.toml
@@ -11,6 +11,7 @@ authors = [
 requires-python = ">=3.12"
 dependencies = [
     "numpy>=2.2.5",
+    "scipy>=1.15.2",
 ]
 
 [build-system]
