+++ src/matrix_lib/__init__.py
@@ -1,8 +1,9 @@
 from .base_matrix import MatrixBase
+from .matrix import FullMatrix, Matrix, SymmetricMatrix
 
 
 def hello() -> str:
     return "Hello from matrix-lib!"
 
 
-__all__ = ["MatrixBase"]
+__all__ = ["MatrixBase", "Matrix", "FullMatrix", "SymmetricMatrix"]
